chore(tokens): Remove final references to old tokens file
