+++ src/components/TokenImage/index.tsx
@@ -5,7 +5,7 @@ import {
   TokenImage as UIKitTokenImage,
   ImageProps,
 } from '@pancakeswap/uikit'
-import tokens from 'config/constants/newTokens'
+import tokens from 'config/constants/tokens'
 import { Token } from '@pancakeswap/sdk'
 
 interface TokenPairImageProps extends Omit<UIKitTokenPairImageProps, 'primarySrc' | 'secondarySrc'> {
