+++ src/config/constants/farms.ts
@@ -1,4 +1,4 @@
-import { serializeTokens } from './newTokens'
+import { serializeTokens } from './tokens'
 import { SerializedFarmConfig } from './types'
 
 const serializedTokens = serializeTokens()
