+++ src/config/constants/priceHelperLps.ts
@@ -1,4 +1,4 @@
-import tokens from './newTokens'
+import tokens from './tokens'
 import { SerializedFarmConfig } from './types'
 
 const priceHelperLps: SerializedFarmConfig[] = [
