+++ src/config/constants/swapWarningTokens.ts
@@ -1,15 +1,10 @@
+import { Token } from '@pancakeswap/sdk'
 import tokens from 'config/constants/tokens'
-import { Address } from './types'
 
 const { bondly, safemoon } = tokens
 
-interface WarningToken {
-  symbol: string
-  address: Address
-}
-
 interface WarningTokenList {
-  [key: string]: WarningToken
+  [key: string]: Token
 }
 
 const SwapWarningTokens = <WarningTokenList>{
