+++ src/hooks/useBUSDPrice.ts
@@ -1,11 +1,11 @@
 import { ChainId, Currency, currencyEquals, JSBI, Price, WETH } from '@pancakeswap/sdk'
 import { useMemo } from 'react'
 import useActiveWeb3React from 'hooks/useActiveWeb3React'
-import { BUSD, CAKE, WBNB } from '../config/constants/tokens'
+import tokens, { mainnetTokens } from 'config/constants/tokens'
 import { PairState, usePairs } from './usePairs'
 import { wrappedCurrency } from '../utils/wrappedCurrency'
 
-const BUSD_MAINNET = BUSD[ChainId.MAINNET]
+const BUSD_MAINNET = mainnetTokens.busd
 
 /**
  * Returns the price in BUSD of the input currency
@@ -72,13 +72,11 @@ export default function useBUSDPrice(currency?: Currency): Price | undefined {
 }
 
 export const useCakeBusdPrice = (): Price | undefined => {
-  const { chainId } = useActiveWeb3React()
-  const currentChaindId = chainId || ChainId.MAINNET
-  const cakeBusdPrice = useBUSDPrice(CAKE[currentChaindId])
+  const cakeBusdPrice = useBUSDPrice(tokens.cake)
   return cakeBusdPrice
 }
 
 export const useBNBBusdPrice = (): Price | undefined => {
-  const bnbBusdPrice = useBUSDPrice(WBNB)
+  const bnbBusdPrice = useBUSDPrice(tokens.wbnb)
   return bnbBusdPrice
 }
