+++ src/utils/addressHelpers.ts
@@ -1,6 +1,6 @@
 import { ChainId } from '@pancakeswap/sdk'
 import addresses from 'config/constants/contracts'
-import tokens from 'config/constants/newTokens'
+import tokens from 'config/constants/tokens'
 import { Address } from 'config/constants/types'
 
 export const getAddress = (address: Address): string => {
