+++ src/views/Pools/components/CakeVaultCard/index.tsx
@@ -4,7 +4,7 @@ import { Box, CardBody, Flex, Text } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
 import { useWeb3React } from '@web3-react/core'
 import ConnectWalletButton from 'components/ConnectWalletButton'
-import tokens from 'config/constants/newTokens'
+import tokens from 'config/constants/tokens'
 import { useCakeVault } from 'state/pools/hooks'
 import { DeserializedPool } from 'state/types'
 import { convertSharesToCake } from 'views/Pools/helpers'
