+++ src/views/Swap/components/SwapWarningModal/index.tsx
@@ -2,7 +2,6 @@ import React, { useEffect } from 'react'
 import styled from 'styled-components'
 import { ModalBody, ModalContainer, Message, ModalHeader, Box, Heading } from '@pancakeswap/uikit'
 import useTheme from 'hooks/useTheme'
-import { getAddress } from 'utils/addressHelpers'
 import { useTranslation } from 'contexts/Localization'
 import { WrappedTokenInfo } from 'state/lists/hooks'
 import SwapWarningTokensConfig from 'config/constants/swapWarningTokens'
@@ -51,11 +50,11 @@ const SwapWarningModal: React.FC<SwapWarningModalProps> = ({ swapCurrency, onDis
   usePreventModalOverlayClick()
 
   const TOKEN_WARNINGS = {
-    [getAddress(SwapWarningTokensConfig.safemoon.address)]: {
+    [SwapWarningTokensConfig.safemoon.address]: {
       symbol: SwapWarningTokensConfig.safemoon.symbol,
       component: <SafemoonWarning />,
     },
-    [getAddress(SwapWarningTokensConfig.bondly.address)]: {
+    [SwapWarningTokensConfig.bondly.address]: {
       symbol: SwapWarningTokensConfig.bondly.symbol,
       component: <BondlyWarning />,
     },
